+++ IOHIDFamily/IOHIDFamilyUnitTests/TestIOHIDEventDriver_DeviceOrientationAngularSensor.m
@@ -54,7 +54,7 @@
     HIDXCTAssertWithParameters ( RETURN_FROM_TEST | COLLECT_TAILSPIN | COLLECT_IOREG,
                                 result == XCTWaiterResultCompleted,
                                 "result:%ld %@",
-                                result,
+                                (long)result,
                                 self.testServiceExpectation);
 
     HIDAngularSensorInputReport01 report;
