+++ IOHIDFamily/IOHIDFamilyUnitTests/TestIOHIDEventDriver_Gyro.m
@@ -52,7 +52,7 @@
     HIDXCTAssertWithParameters ( RETURN_FROM_TEST | COLLECT_TAILSPIN | COLLECT_IOREG,
                                 result == XCTWaiterResultCompleted,
                                 "result:%ld %@",
-                                result,
+                                (long)result,
                                 self.testServiceExpectation);
 
     HIDGyroInputReport01 report;
